requirements.txt
APScheduler==3.6.3
autopep8==1.6.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
google-api-core==2.8.2
google-api-python-client==2.29.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
install==1.3.5
protobuf==4.21.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.9
pyTelegramBotAPI==4.6.1
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.3
tdqm==0.0.1
toml==0.10.2
tornado==6.2
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.11
